sample_project/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
crispy-bootstrap5==2025.6
    # via -r requirements.in
dj-database-url==3.0.1
    # via -r requirements.in
django==5.2.5
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   dj-database-url
    #   django-cors-headers
    #   django-crispy-forms
    #   django-filter
    #   djangorestframework
django-cors-headers==4.7.0
    # via -r requirements.in
django-crispy-forms==2.4
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-filter==25.1
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
dotenv==0.9.9
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
python-dotenv==1.1.1
    # via dotenv
sqlparse==0.5.3
    # via django
